Moved JavaWorkarounds to a new package.
