+++ geode-common/src/main/java/org/apache/geode/internal/lang/JavaWorkarounds.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package org.apache.geode.util.internal;
+package org.apache.geode.internal.lang;
 
 import java.util.Map;
 import java.util.function.Function;
