+++ geode-core/src/jmh/java/org/apache/geode/internal/util/ComputeIfAbsentBenchmark.java
@@ -32,7 +32,7 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 
-import org.apache.geode.util.internal.JavaWorkarounds;
+import org.apache.geode.internal.lang.JavaWorkarounds;
 
 
 /**
