+++ geode-core/src/main/java/org/apache/geode/internal/cache/backup/BackupDefinition.java
@@ -23,7 +23,7 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.geode.cache.DiskStore;
-import org.apache.geode.util.internal.JavaWorkarounds;
+import org.apache.geode.internal.lang.JavaWorkarounds;
 
 class BackupDefinition {
