+++ geode-lucene/src/main/java/org/apache/geode/cache/lucene/FlatFormatSerializer.java
@@ -27,9 +27,9 @@ import org.apache.logging.log4j.Logger;
 import org.apache.lucene.document.Document;
 
 import org.apache.geode.cache.lucene.internal.repository.serializer.SerializerUtil;
+import org.apache.geode.internal.lang.JavaWorkarounds;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 import org.apache.geode.pdx.PdxInstance;
-import org.apache.geode.util.internal.JavaWorkarounds;
 
 /**
  * A built-in {@link LuceneSerializer} to parse user's nested object into a flat format, i.e. a
